POC: generic Test_Simple_Flow
