+++ flow/connectors/postgres/postgres_schema_delta_test.go
@@ -196,23 +196,25 @@ func (s PostgresSchemaDeltaTestSuite) TestAddDropWhitespaceColumnNames() {
 }
 
 func TestPostgresSchemaDeltaTestSuite(t *testing.T) {
-	e2eshared.RunSuite(t, SetupSuite, func(s PostgresSchemaDeltaTestSuite) {
-		teardownTx, err := s.connector.conn.Begin(context.Background())
-		require.NoError(s.t, err)
-		defer func() {
-			err := teardownTx.Rollback(context.Background())
-			if err != pgx.ErrTxClosed {
-				require.NoError(s.t, err)
-			}
-		}()
-		_, err = teardownTx.Exec(context.Background(), fmt.Sprintf("DROP SCHEMA IF EXISTS %s CASCADE",
-			s.schema))
-		require.NoError(s.t, err)
-		err = teardownTx.Commit(context.Background())
-		require.NoError(s.t, err)
+	e2eshared.RunSuite(t, SetupSuite)
+}
 
-		require.NoError(s.t, s.connector.ConnectionActive(context.Background()))
-		require.NoError(s.t, s.connector.Close())
-		require.Error(s.t, s.connector.ConnectionActive(context.Background()))
-	})
+func (s PostgresSchemaDeltaTestSuite) Teardown() {
+	teardownTx, err := s.connector.conn.Begin(context.Background())
+	require.NoError(s.t, err)
+	defer func() {
+		err := teardownTx.Rollback(context.Background())
+		if err != pgx.ErrTxClosed {
+			require.NoError(s.t, err)
+		}
+	}()
+	_, err = teardownTx.Exec(context.Background(), fmt.Sprintf("DROP SCHEMA IF EXISTS %s CASCADE",
+		s.schema))
+	require.NoError(s.t, err)
+	err = teardownTx.Commit(context.Background())
+	require.NoError(s.t, err)
+
+	require.NoError(s.t, s.connector.ConnectionActive(context.Background()))
+	require.NoError(s.t, s.connector.Close())
+	require.Error(s.t, s.connector.ConnectionActive(context.Background()))
 }
